+++ src/app/model/assets/mining.png
