+++ src/app/model/assets/minting.png
