--- src/app/model/components/LoadingSpinner.tsx
@@ -1,9 +1,0 @@
-export default function LoadingSpinner() {
-	return (
-		<div className="loading-spinner">
-			<div className="loading-spinner__circle"></div>
-			<div className="loading-spinner__circle"></div>
-			<div className="loading-spinner__circle"></div>
-		</div>
-	);
-}
